refactor: move tests for moved function
