+++ packages/dependency-graph-integrator/src/pull-requests.ts
@@ -90,9 +90,6 @@ export async function requestTeamReview(
 	}
 }
 
-
-
-
 export async function createPrAndAddToProject(
 	stage: string,
 	repoName: string,
